third_party/WebKit/LayoutTests/fast/hidpi/image-srcset-intrinsic-size-zero.html

<html>
<head>
<script>
    window.targetScaleFactor = 2;
</script>
<script src="resources/srcset-helper.js"></script>
<script src="../../resources/js-test.js"></script>
<script>
    addEventListener("load", function() {
        shouldBe('document.getElementById("foo").clientWidth', '33554432');
    }, false);
</script>
</head>

<body id="body">
    <div>
        This test passes if the image is displayed with infinite dimensions
    </div>
    <img id="foo" srcset="resources/blue-100-px-square.png 0x">
</body>
</html>
