third_party/WebKit/LayoutTests/fast/hidpi/image-srcset-intrinsic-size-zero-expected.txt

PASS successfullyParsed is true

TEST COMPLETE
PASS document.getElementById("foo").clientWidth is 33554432
This test passes if the image is displayed with infinite dimensions
